+++ cmake/CMakePresets.json
@@ -19,7 +19,8 @@
 			"inherits": "default",
 			"cacheVariables": {
 				"CMAKE_C_COMPILER": "clang",
-				"CMAKE_CXX_COMPILER": "clang++"
+				"CMAKE_CXX_COMPILER": "clang++",
+				"CMAKE_EXE_LINKER_FLAGS": "-fuse-ld=lld"
 			}
 		},
 		{
